# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TNC Trading API
  description: |
    API de la plateforme TNC Trading - Tokenisation d'Or Souveraine pour le Burkina Faso.
    
    ## Authentification
    L'API utilise des JWT Bearer tokens. Après login, incluez le token dans le header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - 100 requêtes/minute par IP
    - 1000 requêtes/minute par utilisateur authentifié
    
    ## Codes d'erreur
    Toutes les erreurs suivent le format standard avec un code et message.
  version: 1.0.0
  contact:
    name: TNC Trading Support
    email: support@tnc-trading.com
  license:
    name: Proprietary
servers:
  - url: https://api.tnc-trading.com/api/v1
    description: Production
  - url: https://staging-api.tnc-trading.com/api/v1
    description: Staging
  - url: http://localhost:8787/api/v1
    description: Development
...
paths:
...
  /wallet:
    get:
      tags: [Wallet]
      summary: Obtenir les informations du wallet
      operationId: getWallet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Informations du wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Liste des utilisateurs
      operationId: listUsers
      security:
        - bearerAuth: []
        - cloudflareAccess: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: kycStatus
          in: query
          schema:
            type: string
            enum: [PENDING, SUBMITTED, APPROVED, REJECTED, EXPIRED]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste paginée des utilisateurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cloudflareAccess:
      type: apiKey
      in: header
      name: CF-Access-JWT-Assertion
      description: Cloudflare Access JWT
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # ===== Auth =====
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        phone:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        country:
          type: string
        kycLevel:
          type: string
          enum: [BASIC, STANDARD, VERIFIED]
        kycStatus:
          type: string
          enum: [PENDING, SUBMITTED, APPROVED, REJECTED, EXPIRED]
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        twoFactorEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tokenBalance:
          type: number
          description: Solde en grammes d'or
        cashBalance:
          type: number
          description: Solde en XOF
        tokenValueXof:
          type: number
          description: Valeur des tokens en XOF
        totalValueXof:
          type: number
          description: Valeur totale (tokens + cash)
        averageBuyPrice:
          type: number
          description: Prix moyen d'achat
        profitLoss:
          type: number
          description: Plus/moins-value
        profitLossPercent:
          type: number
        updatedAt:
          type: string
          format: date-time
...